requirements.txt
# Data Processing
pandas
numpy
scikit-learn

# ML & Experimentation
mlflow
xgboost
imbalanced-learn
shap
# Visualization
matplotlib
seaborn
plotly

# Development
jupyter
pytest
black
pylint
pre-commit

# Config
pyyaml
python-dotenv

# API Dependencies
fastapi
uvicorn
python-multipart
pydantic
httpx